Original change: https://googleplex-android-review.googlesource.com/c/platform/system/chre/+/14389479

+++ chpp/transport.c
@@ -331,7 +331,7 @@ static size_t chppConsumeFooter(struct ChppTransportState *context,
                 context->rxHeader.packetCode);
 
       context->rxStatus.lastGoodPacketTimeMs =
-          (uint32_t)(chppGetCurrentTimeNs() / CHPP_NSEC_PER_MSEC);
+          (uint32_t)chppGetCurrentTimeNs() / (uint32_t)CHPP_NSEC_PER_MSEC;
       context->rxStatus.receivedPacketCode = context->rxHeader.packetCode;
       chppRegisterRxAck(context);
 
@@ -1189,7 +1189,7 @@ bool chppRxDataCb(struct ChppTransportState *context, const uint8_t *buf,
   CHPP_LOGD("RX %" PRIuSIZE " bytes: state=%" PRIu8, len,
             context->rxStatus.state);
   context->rxStatus.lastDataTimeMs =
-      (uint32_t)(chppGetCurrentTimeNs() / CHPP_NSEC_PER_MSEC);
+      (uint32_t)chppGetCurrentTimeNs() / (uint32_t)CHPP_NSEC_PER_MSEC;
   context->rxStatus.numTotalDataBytes += len;
 
   size_t consumed = 0;
